[Fix] #560 - delegate 코드 파일 분리
